Broken Checkpoint. Moved most NBallotParser functionality into CandidateRankingParser.
